requirements.txt
flask>=3.0

# Test dependencies
pytest>=7.0
hypothesis>=6.0
pytest-html
allure-pytest
requests
playwright
PyYAML
locust
